# src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: Spring PetClinic
  description: Spring PetClinic Sample Application.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: '1.0'
servers:
  - url: http://localhost:9966/petclinic/api
...
paths:
...
  /specialties:
    get:
      tags:
        - specialties
      operationId: listSpecialties
      summary: Lists specialties
      description: Returns an array of specialty .
      parameters:
        - name: If-None-Match
          in: header
          description: ETag value from a previous response, for conditional GET.
          required: false
          schema:
            type: string
      responses:
        200:
          description: Specialties found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Specialty'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        500:
          description: Server error.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    post:
      tags:
        - specialties
      operationId: addSpecialty
      summary: Create a specialty
      description: Creates a specialty .
      requestBody:
        description: The specialty
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpecialtyFields'
        required: true
      responses:
        201:
          description: Specialty created successfully.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Specialty'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        404:
          description: Specialty not found.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        500:
          description: Server error.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /specialties/{specialtyId}:
    get:
      tags:
        - specialties
      operationId: getSpecialty
      summary: Get a specialty by ID
      description: Returns the specialty or a 404 error.
      parameters:
        - name: specialtyId
          in: path
          description: The ID of the speciality.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 1
        - name: If-None-Match
          in: header
          description: ETag value from a previous response, for conditional GET.
          required: false
          schema:
            type: string
      responses:
        200:
          description: Specialty details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Specialty'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        404:
          description: Specialty not found.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        500:
          description: Server error.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    put:
      tags:
        - specialties
      operationId: updateSpecialty
      summary: Update a specialty by ID
      description: Returns the specialty or a 404 error.
      parameters:
        - name: specialtyId
          in: path
          description: The ID of the specialty.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 1
      requestBody:
        description: The pet
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Specialty'
        required: true
      responses:
        204:
          description: Specialty details found and returned.
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        404:
          description: Specialty not found.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        500:
          description: Server error.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    delete:
      tags:
        - specialties
      operationId: deleteSpecialty
      summary: Delete a specialty by ID
      description: Returns the specialty or a 404 error.
      parameters:
        - name: specialtyId
          in: path
          description: The ID of the specialty.
          required: true
          schema:
            type: integer
            format: int32
            example: 3
      responses:
        200:
          description: Specialty details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Specialty'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        404:
          description: Specialty not found.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        500:
          description: Server error.
          content:
            "application/problem+json":
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
components:
  schemas:
    ProblemDetail:
      title: Problem Detail
      description: The schema for all error responses.
      type: object
      properties:
        type:
          title: Type
          description: Full URL that originated the error response.
          type: string
          format: uri
          example: "http://localhost:9966/petclinic/api/owner"
          readOnly: true
        title:
          title: Title
          description: The short error title.
          type: string
          example: NoResourceFoundException
          readOnly: true
        status:
          title: Status
          description: HTTP status code
          type: integer
          format: int32
          minimum: 400
          maximum: 600
          exclusiveMaximum: true
          example: 500
          readOnly: true
        detail:
          title: Detail
          description: The long error message.
          type: string
          example: "No static resource api/owner."
          readOnly: true
        instance:
          title: Instance
          description: URI identifying the specific occurrence of the problem.
          type: string
          format: uri-reference
          example: "/petclinic/api/v2/owners"
          readOnly: true
        timestamp:
          title: Timestamp
          description: The time the error occurred.
          type: string
          format: date-time
          example: "2024-11-23T13:59:21.3820407Z"
          readOnly: true
        schemaValidationErrors:
          title: Schema validation errors
          description: Validation errors against the OpenAPI schema.
          type: array
          items:
            $ref: '#/components/schemas/ValidationMessage'
      required:
        - type
        - title
        - status
        - detail
        - timestamp
        - schemaValidationErrors
    ValidationMessage:
      title: Validation message
      description: Messages describing a validation error.
      type: object
      properties:
        message:
          title: Message
          description: The validation message.
          type: string
          example: "[Path '/lastName'] Instance type (null) does not match any allowed primitive type (allowed: ['string'])"
          readOnly: true
      required:
        - message
      additionalProperties: true
    SpecialtyFields:
      title: SpecialtyFields
      description: Editable fields of specialty of vets.
      type: object
      properties:
        name:
          title: Name
          description: The name of the specialty.
          type: string
          maxLength: 80
          minLength: 1
          example: radiology
      required:
        - name
    Specialty:
      title: Specialty
      description: Fields of specialty of vets.
      type: object
      properties:
        id:
          title: ID
          description: The ID of the specialty.
          type: integer
          format: int32
          minimum: 1
          maximum: 1000
          example: 1
          readOnly: true
        name:
          title: Name
          description: The name of the specialty.
          type: string
          maxLength: 80
          minLength: 1
          example: radiology
      required:
        - id
        - name
...